refactoring node code
